requirements.txt
google-genai==1.32.0
python-dotenv==1.0.1
pillow==11.3.0
numpy==2.2.6
scikit-learn==1.7.1
streamlit==1.49.1
streamlit-image-comparison==0.0.4